+++ pcbnew/drc/drc_test_provider.h
@@ -119,8 +119,8 @@ protected:
 
     bool isInvisibleText( const BOARD_ITEM* aItem ) const;
 
-    wxString formatMsg( const wxString& aFormatString, const wxString& aSource, int aConstraint,
-                        int aActual );
+    wxString formatMsg( const wxString& aFormatString, const wxString& aSource, double aConstraint,
+                        double aActual );
 
     // List of basic (ie: non-compound) geometry items
     static std::vector<KICAD_T> s_allBasicItems;
